Team08_DontCare/First Round Submission/requirements.txt
Django>=3.2.0
pytesseract>=0.3.8
Pillow>=9.0.0
google-generativeai>=0.2.0
python-dotenv>=0.19.0
openpyxl>=3.0.0
